% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{\%+=\%}
\alias{\%+=\%}
\title{compoundAssignment [`+=`]}
\usage{
t \%+=\% s
}
\arguments{
\item{t}{(`atomic(1)`)\cr
The original variable value.}

\item{s}{(`atomic(1)`)\cr
The value to be added.}
}
\description{
Enables the user the option to use the [`+=`] compound assignment.
}
\examples{
value <- 5
print(value)
value += 7
print(value)

}
